[adapters-only] deduplicate common methods to a trait I
