+++ src/Adapter/Request.php
@@ -10,6 +10,8 @@ use brnc\Symfony1\Message\Obligation\SfWebRequestSubsetInterface;
  */
 class Request implements CommonHeadInterface // TODO implements ServerRequestInterface
 {
+    use CommonAdapterTrait;
+
     /** @var bool[] */
     protected static $contentHeaders = ['CONTENT_LENGTH' => true, 'CONTENT_MD5' => true, 'CONTENT_TYPE' => true];
 
@@ -19,12 +21,12 @@ class Request implements CommonHeadInterface // TODO implements ServerRequestInt
     /** @var \ReflectionProperty */
     protected $reflexivePropertyPathInfoArray;
 
-    /**
-     * @var string[]
-     *
-     * shadow to honour: »[…]preserve the exact case in which headers were originally specified.«
-     */
-    protected $headerNames = [];
+    // /**
+    //  * @var string[]
+    //  *
+    //  * shadow to honour: »[…]preserve the exact case in which headers were originally specified.«
+    //  */
+    // protected $headerNames = [];
 
     /**
      * @var string
@@ -117,32 +119,32 @@ class Request implements CommonHeadInterface // TODO implements ServerRequestInt
         return $value === null? '' : $value;
     }
 
-    /**
-     * @param string          $name
-     * @param string|string[] $value
-     *
-     * @return $this In conflict with PSR-7's immutability paradigm, this method return not a clone but the very same
-     *               instance due to the nature of the underlying adapted symfony object
-     */
-    public function withAddedHeader($name, $value)
-    {
-        if (!$this->hasHeader($name)) {
-            // to preserve the original header name
-            return $this->withHeader($name, $value);
-        }
-
-        $headers = $this->getHeader($name);
-        if (is_array($value)) {
-            $headers = array_merge($headers, $value);
-        }
-        else {
-            $headers[] = $headers;
-        }
-
-        $this->setHeader($name, $headers);
-
-        return $this;
-    }
+    // /**
+    //  * @param string          $name
+    //  * @param string|string[] $value
+    //  *
+    //  * @return $this In conflict with PSR-7's immutability paradigm, this method return not a clone but the very same
+    //  *               instance due to the nature of the underlying adapted symfony object
+    //  */
+    // public function withAddedHeader($name, $value)
+    // {
+    //     if (!$this->hasHeader($name)) {
+    //         // to preserve the original header name
+    //         return $this->withHeader($name, $value);
+    //     }
+    //
+    //     $headers = $this->getHeader($name);
+    //     if (is_array($value)) {
+    //         $headers = array_merge($headers, $value);
+    //     }
+    //     else {
+    //         $headers[] = $headers;
+    //     }
+    //
+    //     $this->setHeader($name, $headers);
+    //
+    //     return $this;
+    // }
 
     /**
      * @param string $name
@@ -154,21 +156,20 @@ class Request implements CommonHeadInterface // TODO implements ServerRequestInt
         return null !== $this->sfWebRequest->getHttpHeader($name);
     }
 
-    /**
-     * @param string          $name
-     * @param string|string[] $value
-     *
-     * @return $this In conflict with PSR-7's immutability paradigm, this method return not a clone but the very same
-     *               instance due to the nature of the underlying adapted symfony object
-     */
-    public function withHeader($name, $value)
-    {
-        $this->headerNames[$this->normalizeHeaderName($name)] = $name;
-
-        $this->setHeader($name, $value);
-
-        return $this;
-    }
+    // /**
+    //  * @param string          $name
+    //  * @param string|string[] $value
+    //  *
+    //  * @return $this In conflict with PSR-7's immutability paradigm, this method return not a clone but the very same
+    //  *               instance due to the nature of the underlying adapted symfony object
+    //  */
+    // public function withHeader($name, $value)
+    // {
+    //     $this->headerNames[$this->normalizeHeaderName($name)] = $name;
+    //     $this->setHeader($name, $value);
+    //
+    //     return $this;
+    // }
 
     /**
      * @param string $name
@@ -273,33 +274,33 @@ class Request implements CommonHeadInterface // TODO implements ServerRequestInt
         $this->reflexivePropertyPathInfoArray->setValue($this->sfWebRequest, $pathInfo);
     }
 
-    /**
-     * @param string $name
-     *
-     * @return string
-     */
-    protected function normalizeHeaderName($name)
-    {
-        return str_replace('_', '-', strtolower($name));
-    }
+    // /**
+    //  * @param string $name
+    //  *
+    //  * @return string
+    //  */
+    // protected function normalizeHeaderName($name)
+    // {
+    //     return str_replace('_', '-', strtolower($name));
+    // }
 
-    /**
-     * Explodes a HTTP header's value to address PSR-7 arrayfied sub-value approach
-     *
-     * N.b. duplication of …
-     *
-     * @see Response::explodeHeaderLine()
-     *
-     * @param string $line
-     *
-     * @return string[]
-     */
-    protected function explodeHeaderLine($line)
-    {
-        return array_map(function($v) {
-            return trim($v, " \t"); // https://tools.ietf.org/html/rfc7230#section-3.2.4
-        }, explode(',', $line));
-    }
+    // /**
+    //  * Explodes a HTTP header's value to address PSR-7 arrayfied sub-value approach
+    //  *
+    //  * N.b. duplication of …
+    //  *
+    //  * @see Response::explodeHeaderLine()
+    //  *
+    //  * @param string $line
+    //  *
+    //  * @return string[]
+    //  */
+    // protected function explodeHeaderLine($line)
+    // {
+    //     return array_map(function($v) {
+    //         return trim($v, " \t"); // https://tools.ietf.org/html/rfc7230#section-3.2.4
+    //     }, explode(',', $line));
+    // }
 
     /**
      * injects a header into symfony's pathInfoArray via setPathInfoArray()'s reflection
@@ -309,10 +310,9 @@ class Request implements CommonHeadInterface // TODO implements ServerRequestInt
      */
     protected function setHeader($name, $value)
     {
-        $keyName = $this->getPathInfoKey($name);
-
+        $keyName                 = $this->getPathInfoKey($name);
         $pathInfoArray           = $this->sfWebRequest->getPathInfoArray();
-        $pathInfoArray[$keyName] = is_array($value)? implode(',', $value) : $value;
+        $pathInfoArray[$keyName] = is_array($value)? implode(',', $value) : $value; // TODO implodeHeaders
         $this->setPathInfoArray($pathInfoArray);
     }
